+++ aiverify-portal/app/inputs/checklists/upload/manual/[checklistId]/components/ChecklistDetail.tsx
