+++ aiverify-portal/app/inputs/checklists/upload/manual/components/GroupDetail.tsx
