+++ aiverify-portal/app/inputs/checklists/upload/manual/components/ProgressSidebar.tsx
